added guard

+++ libft/ft_atoi.c
@@ -6,7 +6,7 @@
 /*   By: 071yoon <071yoon@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/11/10 16:11:29 by 071yoon           #+#    #+#             */
-/*   Updated: 2021/11/11 19:40:33 by 071yoon          ###   ########.fr       */
+/*   Updated: 2021/11/13 17:29:49 by 071yoon          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -22,7 +22,7 @@ static int	is_space(char c)
 int	ft_atoi(const char *str)
 {
 	int			i;
-	long long	result;
+	long double	result;
 	long long	sign;
 
 	i = 0;
@@ -45,5 +45,5 @@ int	ft_atoi(const char *str)
 	else if (result * sign < -2147483648)
 		return (0);
 	else
-		return (result * sign);
+		return ((int)(result * sign));
 }
